requirements.txt
cos-python-sdk-v5>=1.9.30
minio>=7.0.0
pandas>=2.0.0
openpyxl>=3.0.0
requests>=2.25.0
python-dotenv>=1.0.0